initial implementation of an executable server
The goal of this PR is to setup a browser version of Blueprint packaged
as a "drag and drop" that's easy to deploy.

The setup is a bit tricky and might not apply well to every kind of
product made using the Theia framework, but it should work for most
usages.

Add `theia-blueprint-browser` Theia extension that contributes the
required customizations to run in the packaged environment (mostly to
accommodate to the new file layout).

Use a `node`-specific Webpack configuration to bundle the backend along
with the browser-specific customizations.

Use a forked `pkg` package to support forking child processes with
`execArgv` options. The fork is setup as a git submodule.